Remove production.log created when building image
If we don't do this, the log file has incorrect permissions set
and the application can't log into it.

The error is:
Rails Error: Unable to access log file. Please ensure that
/opt/app-root/src/log/production.log exists and is writable (ie, make it
writable for user and group: chmod 0664
/opt/app-root/src/log/production.log). The log level has been raised to
WARN and the output directed to STDERR until the problem is fixed.